+++ desk/deskfun.h
@@ -10,7 +10,7 @@
 #ifndef _DESKFUN_H
 #define _DESKFUN_H
 
-WORD fun_alert(WORD defbut, WORD stnum, WORD *pwtemp);
+WORD fun_alert(WORD defbut, WORD stnum, BYTE *pwtemp);
 void fun_msg(WORD type, WORD w3, WORD w4, WORD w5, WORD w6, WORD w7);
 void fun_rebld(WNODE *pwin);
 WORD fun_mkdir(WNODE *pw_node);
